requirements.txt

# ============================================================================
# 🌧️ REQUIREMENTS.TXT - API + ETL DADOS PLUVIOMÉTRICOS e METEOROLÓGICOS
# ============================================================================
# 
# Sistema completo para sincronização e API REST de dados pluviométricos
# Servidor: 10.50.30.166
# Python: 3.8+
#
# ============================================================================

# ============================================================================
# CORE - Framework Web
# ============================================================================
Flask
Werkzeug

# ============================================================================
# CORS - Permitir requisições cross-origin
# ============================================================================
Flask-CORS

# ============================================================================
# BANCO DE DADOS - PostgreSQL
# ============================================================================
psycopg2-binary
# Alternativa (se psycopg2-binary der problema):
psycopg2

# ============================================================================
# SERVIDOR WSGI PARA PRODUÇÃO
# ============================================================================
gunicorn

# ============================================================================
# VARIÁVEIS DE AMBIENTE
# ============================================================================
python-dotenv

# ============================================================================
# UTILITÁRIOS
# ============================================================================
# Manipulação de datas
python-dateutil

# Requisições HTTP (se precisar fazer requests externos)
requests

# ============================================================================
# MONITORAMENTO E LOGS (OPCIONAL)
# ============================================================================
# Logs estruturados
python-json-logger

# Monitoramento de performance
psutil

# ============================================================================
# SEGURANÇA (OPCIONAL)
# ============================================================================
# Rate limiting para API
Flask-Limiter

# Validação de dados
marshmallow

# ============================================================================
# TESTES (DESENVOLVIMENTO)
# ============================================================================
pytest
pytest-cov
pytest